fix UnitConverter compilation issues

+++ diff-gps/src/UnitConverter.cpp
@@ -42,7 +42,7 @@ vector<long double> UnitConverter::cartToGeo(long double x, long double y, long 
 
 // input: seconds in unix time
 // output: the year as a string
-string UnitConverter::getYear(long long unixTime) {
+string UnitConverter::getYear(long unixTime) {
   ostringstream oss;
   oss << unixTime / (60 * 60 * 24 * 365);
   return oss.str();
@@ -50,7 +50,7 @@ string UnitConverter::getYear(long long unixTime) {
 
 // input: seconds in unix time
 // output: the day of the year as a string (1 is Jan 1, 365 is Dec 31)
-string UnitConverter::getDayOfYear(long long unixTime) {
+string UnitConverter::getDayOfYear(long unixTime) {
   ostringstream oss;
   oss << (unixTime / (60 * 60 * 24)) % 365;
   return oss.str();
